[SYSTEMML-1428] Fixed maxpooling functions for padding > 0
Closes #437.

+++ src/main/java/org/apache/sysml/yarn/ropt/ResourceConfig.java
@@ -101,7 +101,7 @@ public class ResourceConfig
 
 	public long getMaxMRResource()
 	{
-		double val = Collections.max(_mrres);
+		double val = (double) Collections.max(_mrres);
 		return (long)val;
 	}
